Rename module replay_buffer to generic_replay_buffer (#18)
