+++ border/examples/gym/dqn_cartpole_tch.rs
@@ -1,10 +1,10 @@
 use anyhow::Result;
 use border_core::{
-    record::AggregateRecorder,
-    replay_buffer::{
-        SimpleReplayBuffer, SimpleReplayBufferConfig, SimpleStepProcessor,
-        SimpleStepProcessorConfig, BatchBase,
+    generic_replay_buffer::{
+        BatchBase, SimpleReplayBuffer, SimpleReplayBufferConfig, SimpleStepProcessor,
+        SimpleStepProcessorConfig,
     },
+    record::AggregateRecorder,
     Agent, Configurable, DefaultEvaluator, Evaluator as _, Trainer, TrainerConfig,
 };
 use border_py_gym_env::{
